@@ -1,7 +1,7 @@
 ﻿using BinaryQuest.Framework.Core.Data;
+using Duende.IdentityServer.Models;
+using Duende.IdentityServer.Services;
 using IdentityModel;
-using IdentityServer4.Models;
-using IdentityServer4.Services;
 using Microsoft.AspNetCore.Identity;
 using System;
 using System.Collections.Generic;
